requirements.txt
pytz
requests
bs4
Django
djangorestframework
django-cors-headers
pillow
gunicorn